+++ tests/expected/infer_ops.kt
@@ -8,16 +8,16 @@ fun foo() {
     val _c4 = (a / b)
     val _c5 = -(a)
     val d = 2.0
-    val _e1 = (a + d)
-    val _e2 = (a - d)
-    val _e3 = (a * d)
-    val _e4 = (a / d)
+    val _e1 = (a.toDouble() + d)
+    val _e2 = (a.toDouble() - d)
+    val _e3 = (a.toDouble() * d)
+    val _e4 = (a.toDouble() / d)
     val _f = -3.0
     val _g = -(a)
 }
 
 fun add1(x: Byte, y: Byte): Short {
-    return (x + y)
+    return (x + y) as Short
 }
 
 fun add2(x: Short, y: Short): Int {
@@ -25,7 +25,7 @@ fun add2(x: Short, y: Short): Int {
 }
 
 fun add3(x: Int, y: Int): Long {
-    return (x + y)
+    return (x + y) as Long
 }
 
 fun add4(x: Long, y: Long): Long {
@@ -33,7 +33,7 @@ fun add4(x: Long, y: Long): Long {
 }
 
 fun add5(x: UByte, y: UByte): UShort {
-    return (x + y)
+    return (x + y) as UShort
 }
 
 fun add6(x: UShort, y: UShort): UInt {
@@ -41,7 +41,7 @@ fun add6(x: UShort, y: UShort): UInt {
 }
 
 fun add7(x: UInt, y: UInt): ULong {
-    return (x + y)
+    return (x + y) as ULong
 }
 
 fun add8(x: ULong, y: ULong): ULong {
@@ -49,24 +49,24 @@ fun add8(x: ULong, y: ULong): ULong {
 }
 
 fun add9(x: Byte, y: UShort): UInt {
-    return (x + y)
+    return (x.toUShort() + y)
 }
 
 fun sub(x: Byte, y: Byte): Byte {
-    return (x - y)
+    return (x - y) as Byte
 }
 
 fun mul(x: Byte, y: Byte): Short {
-    return (x * y)
+    return (x * y) as Short
 }
 
 fun fadd1(x: Byte, y: Double): Double {
-    return (x + y)
+    return (x.toDouble() + y)
 }
 
 fun show() {
     val rv = fadd1(6, 6.0)
-    assert(rv == 12)
+    assert(rv == 12.toDouble())
     println("OK")
 }
